webfrontend: Split api to blueprint, refactor
