+++ video2commons/frontend/shared.py
@@ -1,0 +1,34 @@
+#!/usr/bin/env python
+# -*- coding: UTF-8 -*-
+#
+# Copyright (C) 2015-2016 Zhuyifei1999
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>
+#
+
+"""video2commons web shared."""
+
+from __future__ import absolute_import
+
+from video2commons.config import redis_pw, redis_host
+from redis import Redis
+
+redisconnection = Redis(host=redis_host, db=3, password=redis_pw)
+
+
+def check_banned():
+    """Check for banned cases."""
+    # Check for WP0 traffic
+
+    return None
